+++ lib/index.d.ts
@@ -1,3 +1,4 @@
+export { default as Validations } from './validations';
 /** Configuration for storeObject */
 export interface StoreObjectConfig<O extends Record<string, any>> {
     /**
@@ -36,24 +37,6 @@ export interface StoreObjectConfig<O extends Record<string, any>> {
     stringify?: (value: any) => string;
 }
 /**
- * A function that can be used to validate that only expected keys are present on an object.
- * Meant to be used in a validate function for `storeObject`
- * @example
- * ```typescript
- * import { storeObject, keyValidation } from 'ls-proxy'
- *
- * const myObj = storeObject(
- *   'myObj',
- *   { foo: 'bar' },
- *   { validate: value => keyValidation(value, ['foo']) },
- * )
- *
- * myObj.foo = 'abc' // no error
- * myObj.bar = 'xyz' // error
- * ```
- */
-export declare const keyValidation: <O extends Record<string, any>>(value: any, requiredKeys: readonly string[]) => boolean | readonly [boolean] | readonly [false, Error];
-/**
  * Store a stringified JSON object in localStorage.
  * This method can use any type that can be serialized.
  * The object stored in localStorage is **not** checked for validity by default,
@@ -83,7 +66,7 @@ export declare const keyValidation: <O extends Record<string, any>>(value: any, 
  * @example
  * ```typescript
  * // Validating that the expected keys exist and are the correct type
- * import { storeObject, keyValidation } from 'ls-proxy'
+ * import { storeObject, validateKeys } from 'ls-proxy'
  *
  * const myObj = storeObject(
  *   'myObj',
@@ -93,7 +76,7 @@ export declare const keyValidation: <O extends Record<string, any>>(value: any, 
  *   },
  *   {
  *     validate(value) {
- *       if (!keyValidation(value, ['someString', 'someNumber'])) return false
+ *       if (!validateKeys(value, ['someString', 'someNumber'])) return false
  *       if (typeof value.someString !== 'string') return false
  *       if (typeof value.someNumber !== 'number') return false
  *       return true