+++ src/index.ts
@@ -1,4 +1,6 @@
-type Keys<O extends Record<string, any>> = keyof O & string
+import type { Keys } from './types'
+
+export { default as Validations } from './validations'
 
 /** Configuration for storeObject */
 export interface StoreObjectConfig<O extends Record<string, any>> {
@@ -39,30 +41,6 @@ export interface StoreObjectConfig<O extends Record<string, any>> {
    */
   stringify?: (value: any) => string
 }
-
-/**
- * A function that can be used to validate that only expected keys are present on an object.
- * Meant to be used in a validate function for `storeObject`
- * @example
- * ```typescript
- * import { storeObject, keyValidation } from 'ls-proxy'
- *
- * const myObj = storeObject(
- *   'myObj',
- *   { foo: 'bar' },
- *   { validate: value => keyValidation(value, ['foo']) },
- * )
- *
- * myObj.foo = 'abc' // no error
- * myObj.bar = 'xyz' // error
- * ```
- */
-export const keyValidation = <O extends Record<string, any>>(
-  value: any,
-  requiredKeys: readonly string[],
-): ReturnType<Required<StoreObjectConfig<O>>['validate']> =>
-  Object.keys(value).every(key => requiredKeys.includes(key)) &&
-  requiredKeys.every(key => key in value)
 
 const defaultStoreObjectConfig = <O extends Record<string, any>>({
   checkGets,
@@ -110,7 +88,7 @@ const defaultStoreObjectConfig = <O extends Record<string, any>>({
  * @example
  * ```typescript
  * // Validating that the expected keys exist and are the correct type
- * import { storeObject, keyValidation } from 'ls-proxy'
+ * import { storeObject, validateKeys } from 'ls-proxy'
  *
  * const myObj = storeObject(
  *   'myObj',
@@ -120,7 +98,7 @@ const defaultStoreObjectConfig = <O extends Record<string, any>>({
  *   },
  *   {
  *     validate(value) {
- *       if (!keyValidation(value, ['someString', 'someNumber'])) return false
+ *       if (!validateKeys(value, ['someString', 'someNumber'])) return false
  *       if (typeof value.someString !== 'string') return false
  *       if (typeof value.someNumber !== 'number') return false
  *       return true