+++ src/validations.ts
@@ -1,0 +1,66 @@
+import type { StoreObjectConfig } from '.'
+import type { Keys } from './types'
+
+/** Validations meant to be used with `storeObject`'s validate function */
+namespace Validations {
+  /**
+   * Validate that only expected keys are present on an object
+   *
+   * @example
+   * ```typescript
+   * import { storeObject, Validations } from 'ls-proxy'
+   *
+   * const myObj = storeObject(
+   *   'myObj',
+   *   { foo: 'bar' },
+   *   { validate: value => Validations.keys(value, ['foo']) },
+   * )
+   *
+   * myObj.foo = 'abc' // no error
+   * myObj.bar = 'xyz' // error
+   * ```
+   */
+  export const keys = <O extends Record<string, any>>(
+    value: Readonly<any>,
+    requiredKeys: readonly string[],
+  ): ReturnType<Required<StoreObjectConfig<O>>['validate']> =>
+    Object.keys(value).every(key => requiredKeys.includes(key)) &&
+    requiredKeys.every(key => key in value)
+
+  /**
+   * Validate that the types passed for an object are expected
+   *
+   * @param value The unknown value to validate types of
+   * @param typesMap A map of expected keys for an object to expected types, checked like `typeof value[key] === typesMap[key]`
+   * @example
+   * ```typescript
+   * import { storeObject, Validations } from 'ls-proxy'
+   *
+   * const typesMap = {
+   *   onlyString: 'string',
+   *   onlyNumber: 'number',
+   * }
+   *
+   * const runtimeCheckedTypes = storeObject(
+   *   'runtimeCheckedTypes',
+   *   {
+   *     onlyString: 'abc',
+   *     onlyNumber: 123,
+   *   },
+   *   { validate: value => Validations.types(value, typesMap) },
+   * )
+   *
+   * runtimeCheckedTypes.onlyString = 'xyz' // Succeeds
+   * runtimeCheckedTypes.onlyNumber = 'abc' // Fails
+   * ```
+   */
+  export const types = <O extends Record<string, any>>(
+    value: Readonly<any>,
+    typesMap: Record<Keys<O>, string>,
+  ) =>
+    Object.entries(value).every(
+      ([key, value]) => typeof value === typesMap[key],
+    )
+}
+
+export default Validations